#830 Changes in the original implementation to speed up the synchronization process
These are the changes we made in the code:

We do bulk updates whenever is possible. We load and process as many changes as possible and we send them to Solr in a bulk.
We avoid "updates", we treat updates as upserts. We treat mongo updates as mongo inserts. Why? because we do not change anything directly into Solr, we only use Solr as a query backend. Therefor, we can infer that the documents we want to update are always there, we do not need to run the two-step process to validate that. Doing so, we can send all the updates in a bulk operation.
We use a high auto-commit-interval, around 20 minutes or so, to avoid flushing changes every time a document (or a bulk of documents) is send to Solr. We need to take into account that we make millions of changes in mongo every time we executed an ETL, so this will have a huge impact.